requirements.txt
streamlit
black
isort
plotly_express